requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
psycopg[binary]==3.3.3
python-dotenv==1.2.2
reportlab==4.4.10
openpyxl==3.1.5
celery==5.6.3
redis==7.4.0
django-redis==6.0.0
django-cors-headers==4.9.0
pytest==9.0.2
pytest-django==4.12.0
